Cleanup Move templater to it's own file Set usage func by default in commands Use nested templates Spacing fixes
